pyproject.toml
# ABOUTME: Project configuration for the A2A-MCP Framework boilerplate
# ABOUTME: Defines dependencies, scripts, and build configuration for multi-agent systems

[project]
name = "a2a-mcp-framework"
version = "0.1.0"
description = "A2A-MCP Framework - Production-ready foundation for multi-agent systems"
readme = "README.md"
requires-python = ">=3.12"
keywords = ["mcp", "agents", "ai", "framework", "a2a", "multi-agent"]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # A2A SDK (Critical for A2A protocol)
    "a2a-sdk>=0.2.9",
    
    # Core MCP Dependencies
    "mcp[cli]>=1.5.0",
    "fastmcp>=1.0",
    
    # Google AI Platform Dependencies (for StandardizedAgentBase)
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.91.0",
    "google-generativeai>=0.8.5",
    
    # LangChain & LangGraph Dependencies (for MasterOrchestratorTemplate)
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.4.1",
    
    # Data Processing & Analysis (for finance agents)
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "networkx>=3.4.2",
    
    # CLI Framework
    "click>=8.1.8",
    
    # Core Python Dependencies
    "httpx>=0.28.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",
    
    # Logging and Utils
    "structlog>=23.1.0",
    "pyyaml>=6.0",
    
    # Async Support
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
# A2A MCP Framework entry points
a2a-mcp = "a2a_mcp:main"
launch-system = "launch.launch_system:main"
start-mcp = "a2a_mcp.mcp.server:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/a2a-mcp-framework"
"Bug Tracker" = "https://github.com/yourusername/a2a-mcp-framework/issues"
"Documentation" = "https://github.com/yourusername/a2a-mcp-framework/blob/main/README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/a2a_mcp/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/agent_cards",
    "/configs",
    "/docs",
    "/examples",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/a2a_mcp", "src/video_generator"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["src/a2a_mcp"]

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
]
ignore = [
    "E501", # line too long
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true